requirements.txt
aiosqlite==0.20.0
base58==2.1.1
loguru==0.7.2
PyNaCl==1.5.0
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2023.3.post1
Requests==2.32.3
solana==0.34.2
websockets==11.0.3
